Revert "chore: update delete application to POST"
This reverts commit 7b73afdbfbfc6363192050dc9297d1289748c58e.
