+++ src/app/api/applications/[id]/route.ts
@@ -2,7 +2,10 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-import { retrieveApplication } from "@/services/daam/index.server";
+import {
+  deleteApplication,
+  retrieveApplication,
+} from "@/services/daam/index.server";
 import { getServerSession } from "next-auth";
 import { NextResponse } from "next/server";
 import { ExtendedSession } from "../../auth/auth.types";
@@ -23,6 +26,26 @@ export async function GET(
   try {
     const { data: application } = await retrieveApplication(id, session);
     return NextResponse.json(application, { status: 200 });
+  } catch (error) {
+    return handleErrorResponse(error);
+  }
+}
+
+export async function DELETE(
+  request: Request,
+  params: { params: { id: string } }
+) {
+  const session: ExtendedSession | null = await getServerSession(authOptions);
+  const { id } = params.params;
+
+  if (!session) {
+    return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+  }
+
+  try {
+    await deleteApplication(id, session);
+
+    return NextResponse.json({ status: 200 });
   } catch (error) {
     return handleErrorResponse(error);
   }
